# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/capacha/verify": {
            "get": {
                "tags": [
                    "登录业务接口"
                ],
                "summary": "验证验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "capachaId",
                        "name": "capachaId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "capachaVal",
                        "name": "capachaVal",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
...
        "/forgetPassword/modifyPassword": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录业务接口"
                ],
                "summary": "密码修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.ModifyPasswordForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
        "/forgetPassword/verifyEmailCode": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录业务接口"
                ],
                "summary": "忘记密码-验证验证码",
                "parameters": [
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.VerifyEmailCodeForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录业务接口"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录业务接口"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.RegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
...
        "/sendEmailCode": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录业务接口"
                ],
                "summary": "发送邮件验证码",
                "parameters": [
                    {
                        "description": "发送参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.SendCodeForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.H": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "utils.LoginForm": {
            "type": "object",
            "required": [
                "password",
                "repassword",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "utils.ModifyPasswordForm": {
            "type": "object",
            "required": [
                "password",
                "repassword",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "utils.RegisterForm": {
            "type": "object",
            "required": [
                "code",
                "email",
                "password",
                "repassword",
                "username"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "repassword": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "utils.SendCodeForm": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "utils.VerifyEmailCodeForm": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
    }
}